requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-secret-manager>=2.16.0
requests>=2.31.0
python-dotenv>=1.0.0
